removed hook, split to functions
